+++ include/httc/server.hpp
@@ -8,8 +8,8 @@
 namespace httc {
 
 void bind_and_listen(
-    std::string_view addr, unsigned int port, std::shared_ptr<Router> router,
-    asio::io_context& io_ctx, ServerConfig config = {}
+    std::string_view addr, uint16_t port, std::shared_ptr<Router> router, asio::io_context& io_ctx,
+    ServerConfig config = {}
 );
 
 }
